requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filetype==1.2.0
fonttools==4.61.1
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-vision==3.12.1
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.7
kiwisolver==1.4.9
matplotlib==3.10.8
mysqlclient==2.2.7
numpy==2.4.1
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
packaging==26.0
pi_heif==1.2.0
pillow==12.1.0
pillow-avif-plugin==1.5.5
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.13
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
